refactor(app-settings): isolate authorized folders panel
Move the authorized-folder config flow into a dedicated settings owner so the app shell and dashboard stop carrying state they do not own while preserving the existing workspace/server behavior.
