+++ apps/app/src/app/app.tsx
@@ -172,7 +172,6 @@ import {
   openworkServerInfo,
   orchestratorStatus,
   opencodeRouterInfo,
-  pickDirectory,
   setWindowDecorations,
   type OrchestratorStatus,
   type OpenworkServerInfo,
@@ -2055,70 +2054,6 @@ export default function App() {
     return typeof compaction.auto === "boolean" ? compaction.auto : null;
   };
 
-  const normalizeAuthorizedFolderPath = (input: string | null | undefined) => {
-    const trimmed = (input ?? "").trim();
-    if (!trimmed) return "";
-    const withoutWildcard = trimmed.replace(/[\\/]\*+$/, "");
-    return normalizeDirectoryQueryPath(withoutWildcard);
-  };
-
-  const authorizedFolderToExternalDirectoryKey = (folder: string) => {
-    const normalized = normalizeAuthorizedFolderPath(folder);
-    if (!normalized) return "";
-    return normalized === "/" ? "/*" : `${normalized}/*`;
-  };
-
-  const externalDirectoryKeyToAuthorizedFolder = (key: string, value: unknown) => {
-    if (value !== "allow") return null;
-    const trimmed = key.trim();
-    if (!trimmed) return null;
-    if (trimmed === "/*") return "/";
-    if (!trimmed.endsWith("/*")) return null;
-    return normalizeAuthorizedFolderPath(trimmed.slice(0, -2));
-  };
-
-  const readAuthorizedFoldersFromConfig = (opencodeConfig: Record<string, unknown>) => {
-    const permission = ensureRecord(opencodeConfig.permission);
-    const externalDirectory = ensureRecord(permission.external_directory);
-    const folders: string[] = [];
-    const hiddenEntries: Record<string, unknown> = {};
-    const seen = new Set<string>();
-
-    for (const [key, value] of Object.entries(externalDirectory)) {
-      const folder = externalDirectoryKeyToAuthorizedFolder(key, value);
-      if (!folder) {
-        hiddenEntries[key] = value;
-        continue;
-      }
-      if (seen.has(folder)) continue;
-      seen.add(folder);
-      folders.push(folder);
-    }
-
-    return { folders, hiddenEntries };
-  };
-
-  const buildAuthorizedFoldersStatus = (preservedCount: number, action?: string) => {
-    const preservedLabel =
-      preservedCount > 0
-        ? `Preserving ${preservedCount} non-folder permission ${preservedCount === 1 ? "entry" : "entries"}.`
-        : null;
-    if (action && preservedLabel) return `${action} ${preservedLabel}`;
-    return action ?? preservedLabel;
-  };
-
-  const mergeAuthorizedFoldersIntoExternalDirectory = (
-    folders: string[],
-    hiddenEntries: Record<string, unknown>,
-  ): Record<string, unknown> | undefined => {
-    const next: Record<string, unknown> = { ...hiddenEntries };
-    for (const folder of folders) {
-      const key = authorizedFolderToExternalDirectoryKey(folder);
-      if (!key) continue;
-      next[key] = "allow";
-    }
-    return Object.keys(next).length ? next : undefined;
-  };
   const [modelPickerOpen, setModelPickerOpen] = createSignal(false);
   const [modelPickerTarget, setModelPickerTarget] = createSignal<
     "session" | "default"
@@ -2147,14 +2082,6 @@ export default function App() {
     setAutoCompactContext((value) => !value);
     setAutoCompactContextDirty(true);
   };
-  const [authorizedFolders, setAuthorizedFolders] = createSignal<string[]>([]);
-  const [authorizedFolderDraft, setAuthorizedFolderDraft] = createSignal("");
-  const [, setAuthorizedFolderHiddenEntries] = createSignal<Record<string, unknown>>({});
-  const [authorizedFoldersLoading, setAuthorizedFoldersLoading] = createSignal(false);
-  const [authorizedFoldersSaving, setAuthorizedFoldersSaving] = createSignal(false);
-  const [authorizedFoldersStatus, setAuthorizedFoldersStatus] = createSignal<string | null>(null);
-  const [authorizedFoldersError, setAuthorizedFoldersError] = createSignal<string | null>(null);
-
   const resolveCodexReasoningEffort = (modelID: string, variant: string | null) => {
     if (!modelID.trim().toLowerCase().includes("codex")) return undefined;
     const normalized = normalizeModelBehaviorValue(variant);
@@ -2933,18 +2860,6 @@ export default function App() {
       openworkServerReady() &&
       openworkServerWorkspaceReady() &&
       (resolvedOpenworkCapabilities()?.plugins?.write ?? false),
-  );
-  const openworkServerCanReadConfig = createMemo(
-    () =>
-      openworkServerReady() &&
-      openworkServerWorkspaceReady() &&
-      (resolvedOpenworkCapabilities()?.config?.read ?? false),
-  );
-  const openworkServerCanWriteConfig = createMemo(
-    () =>
-      openworkServerReady() &&
-      openworkServerWorkspaceReady() &&
-      (resolvedOpenworkCapabilities()?.config?.write ?? false),
   );
   const devtoolsCapabilities = createMemo(() => openworkServerCapabilities());
 
@@ -5558,166 +5473,6 @@ export default function App() {
   });
 
   createEffect(() => {
-    const openworkClient = openworkServerClient();
-    const openworkWorkspaceId = runtimeWorkspaceId();
-    const canReadConfig = openworkServerCanReadConfig();
-
-    if (!openworkClient || !openworkWorkspaceId || !canReadConfig) {
-      setAuthorizedFolders([]);
-      setAuthorizedFolderDraft("");
-      setAuthorizedFolderHiddenEntries({});
-      setAuthorizedFoldersLoading(false);
-      setAuthorizedFoldersStatus(null);
-      setAuthorizedFoldersError(null);
-      return;
-    }
-
-    let cancelled = false;
-    setAuthorizedFolderDraft("");
-    setAuthorizedFoldersLoading(true);
-    setAuthorizedFoldersError(null);
-    setAuthorizedFoldersStatus(null);
-
-    const loadAuthorizedFolders = async () => {
-      try {
-        const config = await openworkClient.getConfig(openworkWorkspaceId);
-        if (cancelled) return;
-        const next = readAuthorizedFoldersFromConfig(ensureRecord(config.opencode));
-        setAuthorizedFolders(next.folders);
-        setAuthorizedFolderHiddenEntries(next.hiddenEntries);
-        setAuthorizedFoldersStatus(
-          buildAuthorizedFoldersStatus(Object.keys(next.hiddenEntries).length),
-        );
-      } catch (error) {
-        if (cancelled) return;
-        const message = error instanceof Error ? error.message : safeStringify(error);
-        setAuthorizedFolders([]);
-        setAuthorizedFolderHiddenEntries({});
-        setAuthorizedFoldersError(message);
-      } finally {
-        if (!cancelled) {
-          setAuthorizedFoldersLoading(false);
-        }
-      }
-    };
-
-    void loadAuthorizedFolders();
-
-    onCleanup(() => {
-      cancelled = true;
-    });
-  });
-
-  const persistAuthorizedFolders = async (nextFolders: string[]) => {
-    const openworkClient = openworkServerClient();
-    const openworkWorkspaceId = runtimeWorkspaceId();
-    if (!openworkClient || !openworkWorkspaceId || !openworkServerCanWriteConfig()) {
-      setAuthorizedFoldersError(
-        "A writable OpenWork server workspace is required to update authorized folders.",
-      );
-      return false;
-    }
-
-    setAuthorizedFoldersSaving(true);
-    setAuthorizedFoldersError(null);
-    setAuthorizedFoldersStatus("Saving authorized folders...");
-
-    try {
-      const currentConfig = await openworkClient.getConfig(openworkWorkspaceId);
-      const currentAuthorizedFolders = readAuthorizedFoldersFromConfig(
-        ensureRecord(currentConfig.opencode),
-      );
-      const nextExternalDirectory = mergeAuthorizedFoldersIntoExternalDirectory(
-        nextFolders,
-        currentAuthorizedFolders.hiddenEntries,
-      );
-
-      await openworkClient.patchConfig(openworkWorkspaceId, {
-        opencode: {
-          permission: {
-            external_directory: nextExternalDirectory,
-          },
-        },
-      });
-      setAuthorizedFolders(nextFolders);
-      setAuthorizedFolderHiddenEntries(currentAuthorizedFolders.hiddenEntries);
-      setAuthorizedFoldersStatus(
-        buildAuthorizedFoldersStatus(
-          Object.keys(currentAuthorizedFolders.hiddenEntries).length,
-          "Authorized folders updated.",
-        ),
-      );
-      markOpencodeConfigReloadRequired();
-      return true;
-    } catch (error) {
-      const message = error instanceof Error ? error.message : safeStringify(error);
-      setAuthorizedFoldersError(message);
-      setAuthorizedFoldersStatus(null);
-      return false;
-    } finally {
-      setAuthorizedFoldersSaving(false);
-    }
-  };
-
-  const addAuthorizedFolder = async () => {
-    const normalized = normalizeAuthorizedFolderPath(authorizedFolderDraft());
-    const workspaceRoot = normalizeAuthorizedFolderPath(workspaceStore.selectedWorkspaceRoot().trim());
-    if (!normalized) return;
-    if (workspaceRoot && normalized === workspaceRoot) {
-      setAuthorizedFolderDraft("");
-      setAuthorizedFoldersStatus("Workspace root is already available.");
-      setAuthorizedFoldersError(null);
-      return;
-    }
-    if (authorizedFolders().includes(normalized)) {
-      setAuthorizedFolderDraft("");
-      setAuthorizedFoldersStatus("Folder is already authorized.");
-      setAuthorizedFoldersError(null);
-      return;
-    }
-
-    const ok = await persistAuthorizedFolders([...authorizedFolders(), normalized]);
-    if (ok) {
-      setAuthorizedFolderDraft("");
-    }
-  };
-
-  const removeAuthorizedFolder = async (folder: string) => {
-    const nextFolders = authorizedFolders().filter((entry) => entry !== folder);
-    await persistAuthorizedFolders(nextFolders);
-  };
-
-  const pickAuthorizedFolder = async () => {
-    if (!isTauriRuntime()) return;
-    try {
-      const selection = await pickDirectory({ title: t("onboarding.authorize_folder", currentLocale()) });
-      const folder = typeof selection === "string" ? selection : Array.isArray(selection) ? selection[0] : null;
-      const normalized = normalizeAuthorizedFolderPath(folder);
-      const workspaceRoot = normalizeAuthorizedFolderPath(workspaceStore.selectedWorkspaceRoot().trim());
-      if (!normalized) return;
-      setAuthorizedFolderDraft(normalized);
-      if (workspaceRoot && normalized === workspaceRoot) {
-        setAuthorizedFolderDraft("");
-        setAuthorizedFoldersStatus("Workspace root is already available.");
-        setAuthorizedFoldersError(null);
-        return;
-      }
-      if (authorizedFolders().includes(normalized)) {
-        setAuthorizedFoldersStatus("Folder is already authorized.");
-        setAuthorizedFoldersError(null);
-        return;
-      }
-      const ok = await persistAuthorizedFolders([...authorizedFolders(), normalized]);
-      if (ok) {
-        setAuthorizedFolderDraft("");
-      }
-    } catch (error) {
-      const message = error instanceof Error ? error.message : safeStringify(error);
-      setAuthorizedFoldersError(message);
-    }
-  };
-
-  createEffect(() => {
     if (typeof window === "undefined") return;
     const workspaceId = workspaceStore.selectedWorkspaceId();
     if (!workspaceId) return;
@@ -6539,27 +6294,7 @@ export default function App() {
       cleanupOpenworkDockerContainers,
       dockerCleanupBusy: dockerCleanupBusy(),
       dockerCleanupResult: dockerCleanupResult(),
-      authorizedFolders: authorizedFolders(),
-      authorizedFolderDraft: authorizedFolderDraft(),
-      setAuthorizedFolderDraft,
-      authorizedFoldersLoading: authorizedFoldersLoading(),
-      authorizedFoldersSaving: authorizedFoldersSaving(),
-      authorizedFoldersError: authorizedFoldersError(),
-      authorizedFoldersStatus: authorizedFoldersStatus(),
-      authorizedFoldersAvailable: openworkServerCanReadConfig(),
-      authorizedFoldersEditable: openworkServerCanWriteConfig(),
-      authorizedFoldersHint: !openworkServerReady()
-        ? "OpenWork server is disconnected."
-        : !openworkServerWorkspaceReady()
-          ? "No active server workspace is selected."
-          : !openworkServerCanReadConfig()
-            ? "OpenWork server config access is unavailable for this workspace."
-            : !openworkServerCanWriteConfig()
-              ? "OpenWork server is connected read-only for workspace config."
-              : null,
-      addAuthorizedFolder,
-      pickAuthorizedFolder,
-      removeAuthorizedFolder,
+      markOpencodeConfigReloadRequired,
       resetAppConfigDefaults,
       notionStatus: notionStatus(),
       notionStatusDetail: notionStatusDetail(),
